Fix: even only for $targets, -Werrror is a bad idea
... baked into the image, I mean. GHC doesn't (yet?) have -Wno-error,
so undoing -Werror becomes a thick hassle when compiling a package you
can't be bothered to fix warnings of.

worse yet, all the -Wno-warn=unused-imports, -Wno-warn=unused-local-binds, ...
invalidate Stack's snapshot ID -- and thus cause a full rebuild of all
dependency packages :/ quite plainly, this sucks.

The behavior can be recovered with --ghc-options=-Werror, or editing the
cabal-file of target package.